requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
MouseInfo==0.1.3
pillow==12.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
pywin32==311
waitress==3.0.2
Werkzeug==3.1.5
